Refactor redux usage
