+++ src/js/model/ScreenType.ts
@@ -3,7 +3,7 @@ export type ScreenType = "SMALL" | "MEDIUM" | "LARGE";
 
 export namespace ScreenType {
 
-    export function getWidthPoint(screenType: ScreenType): number {
+    function getWidthPoint(screenType: ScreenType): number {
         switch (screenType) {
             case "SMALL": return 600;
             case "MEDIUM": return 960;
